LayoutTests/http/tests/site-isolation/permissions-policy-expected.txt

CONSOLE MESSAGE: Permission policy 'SyncXHR' check failed for document with origin 'http://localhost:8000'.
Verifies permissions policy is set correctly on cross-site frames

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is 'exception'
PASS event.origin is 'http://localhost:8000'
PASS event.data is 'pass'
PASS event.origin is 'http://localhost:8000'
PASS successfullyParsed is true

TEST COMPLETE
